requirements.txt
fastapi
uvicorn
click
python-dotenv
nuclia
streamlit>=1.28
requests>=2.31
pytest>=7.0
